requirements.txt
azure-cosmos==3.1.1
Flask==1.1.1
pytest==6.0.1
pytest-black==0.3.7
pytest-cov==2.7.1
pyOpenSSL==19.1.0
